Split Ready in log from next.config.js loading
